Update license and move polygon tool.
