+++ polygon_coverage_ros/package.xml
@@ -8,7 +8,7 @@
   <author>Rik Bähnemann</author>
   <maintainer email="brik@ethz.ch">Rik Bähnemann</maintainer>
 
-  <license>BSD</license>
+  <license>GPL</license>
 
   <buildtool_depend>catkin</buildtool_depend>
   <buildtool_depend>catkin_simple</buildtool_depend>
