+++ rviz_polygon_tool/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 2.8.3)
-project(fm_rviz_polygon_tool)
+project(rviz_polygon_tool)
 
 add_definitions(-std=c++11)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall")
@@ -29,7 +29,7 @@ endif()
 #############
 # Header files as explained in https://answers.ros.org/question/215487/could-not-load-panel-in-rviz-pluginlibfactory-the-plugin-for-class/
 set(HEADER_FILES
-  include/fm_rviz_polygon_tool/polygon_tool.h
+  include/rviz_polygon_tool/polygon_tool.h
 )
 set(SRC_FILES
   src/polygon_tool.cc
