+++ rviz_polygon_tool/package.xml
@@ -1,15 +1,16 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>fm_rviz_polygon_tool</name>
+  <name>rviz_polygon_tool</name>
   <version>0.0.0</version>
   <description>
     RVIZ tool to create polygons.
   </description>
 
   <maintainer email="brik@ethz.ch">Rik Bähnemann</maintainer>
-  <maintainer email="tmarvel@student.ethz.ch">Timothy Marvel</maintainer>
+  <author email="brik@ethz.ch">Rik Bähnemann</author>
+  <author email="tmarvel@student.ethz.ch">Timothy Marvel</author>
 
-  <license>TBD</license>
+  <license>GPL</license>
 
   <buildtool_depend>catkin</buildtool_depend>
   <buildtool_depend>catkin_simple</buildtool_depend>
