+++ rviz_polygon_tool/plugin_description.xml
@@ -1,6 +1,6 @@
-<library path="lib/libfm_rviz_polygon_tool">
-  <class name="fm_rviz_polygon_tool/PolygonSelection"
-         type="mav_polygon_tool::PolygonTool"
+<library path="lib/librviz_polygon_tool">
+  <class name="rviz_polygon_tool/PolygonSelection"
+         type="rviz_polygon_tool::PolygonTool"
          base_class_type="rviz::Tool">
     <description>
       Tool for drawing polygons in RVIZ.
